{"types":"C","runs":11}

+++ Problem15.java
@@ -2,8 +2,8 @@ import acm.program.*;
 
 public class Problem15 extends ConsoleProgram {
 	public void run() {
-		double x = readInt("Enter the first number: ");
-		double y = readInt("Enter the second number: ");
+		int x = readInt("Enter the first number: ");
+		int y = readInt("Enter the second number: ");
 		println("The arithmetic mean of these numbers is: " + (double) (( x + y ) / 2));
 	}
 }
